Make unlimited lag frames
If the client lags, the ingame time may suffer.
Verification of such games might pose a problem, as IGT is challenged
against real time (to close a way for cheaters to slow down the game).
Put the burden of lag on the player instead, no matter how much is it.